+++ functions/src/controllers/scheduled/main.ts
@@ -7,8 +7,14 @@ const main = async (event: ScheduledEvent) => {
   const timestamp = DateTime.fromISO(event.scheduleTime);
 
   await Promise.allSettled([
-    crontab("0 7 * * *", () => import("./notifyDayduty.js"))(timestamp),
-    crontab("0 7 * * *", () => import("./notifyEvent.js"))(timestamp),
+    crontab(
+      "0 7 * * *",
+      () => import("../../services/notify_datduty/index.js"),
+    )(timestamp),
+    crontab(
+      "0 7 * * *",
+      () => import("../../services/notify_events/index.js"),
+    )(timestamp),
   ]);
 };
 export default main;
