+++ functions/src/services/notify_datduty/index.ts
@@ -1,0 +1,1 @@
+export { notifyDayDuty as default } from "./notify_dayduty.js";
