+++ functions/src/services/notify_datduty/notify_dayduty_per_group.ts
@@ -2,19 +2,22 @@ import {
   type ChatPostMessageArguments,
   WebClient as SlackWebClient,
 } from "@slack/web-api";
-import { getFirestore } from "firebase-admin/firestore";
 import type { DateTime } from "luxon";
-import { firebaseApp } from "~/client/firebaseApp.js";
 import { fetchSecret } from "~/services/fetch_secret.js";
-import type { CrontabHandler } from "../../core/crontab.js";
 import { getDayduty } from "../../core/dayduty/getDayduty.js";
-import { getTeacher } from "../../services/classmates/getTeacher.js";
+import { getTeacher } from "../classmates/getTeacher.js";
 
 type Group = {
   id: string;
 };
 
-const notifyDayDutyPerGroup = async (group: Group, timestamp: DateTime) => {
+/**
+ * 日直の通知
+ */
+export const notifyDayDutyPerGroup = async (
+  group: Group,
+  timestamp: DateTime,
+) => {
   const slackToken = await fetchSecret(`group_${group.id}-slack_token`);
   const slackClient = new SlackWebClient(slackToken);
 
@@ -37,20 +40,3 @@ const notifyDayDutyPerGroup = async (group: Group, timestamp: DateTime) => {
     slackClient.chat.postMessage(options);
   }
 };
-
-/**
- * 日直の通知
- */
-const notifyDayDuty: CrontabHandler = async (timestamp) => {
-  const firestore = getFirestore(firebaseApp);
-
-  const groupsSnapshot = await firestore.collection("groups").get();
-  const groups = groupsSnapshot.docs.map((doc) => ({
-    id: doc.id,
-  }));
-
-  await Promise.all(
-    groups.map((group) => notifyDayDutyPerGroup(group, timestamp)),
-  );
-};
-export default notifyDayDuty;
