+++ functions/src/services/notify_events/index.ts
@@ -1,0 +1,1 @@
+export { notifyEvent as default } from "./notify_event.js";
